refactor(dropdownMenu): rename handleClick to onClick
