+++ src/components/inputs/Select/DropdownMenu/DropdownMenu.stories.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import { DropdownMenu } from "../index";
+import { DropdownMenu } from "./index";
 
-import { props, parameters } from "../props";
+import { props, parameters } from "./props";
 
 const story = {
   title: "Inputs/DropdownMenu",
@@ -14,9 +14,9 @@ const story = {
 const Default = (args) => <DropdownMenu {...args} />;
 Default.args = {
   options: [
-    { id: "1", label: "Item 1", disabled: true, isSelected: false },
-    { id: "2", label: "Item 2", disabled: false, isSelected: false },
-    { id: "3", label: "Item 3", disabled: false, isSelected: false },
+    { id: "1", label: "Item 1", disabled: true, selected: false },
+    { id: "2", label: "Item 2", disabled: false, selected: false },
+    { id: "3", label: "Item 3", disabled: false, selected: false },
   ],
   isOpenOptions: true,
 };
